backend/requirements.txt
Flask
mysql_connector-python
